fix(fe): 로그인 여부에 따라 페이지 이동
공통적으로 로그인이 필요한 페이지는 with-authentication 함수를 통해  로그인 하지 않았을 경우 로그인 페이지 이동
