+++ web/pages/profile.js
@@ -3,14 +3,17 @@ import React from 'react';
 import Profile from '../components/Profile';
 import * as GS from '../components/GlobalStyle';
 import BackButton from '../components/BackButton';
+import { withAuthentication } from '../utils/with-authentication';
 
-const ProfilePage = () => (
-  <GS.FlexRowCenterWrap>
-    <BackButton />
-    <GS.TinyBoard>
-      <Profile />
-    </GS.TinyBoard>
-  </GS.FlexRowCenterWrap>
-);
+const ProfilePage = ({ name, sub_email, email }) => {
+  return (
+    <GS.FlexRowCenterWrap>
+      <BackButton />
+      <GS.TinyBoard>
+        <Profile name={name} sub_email={sub_email} email={email} />
+      </GS.TinyBoard>
+    </GS.FlexRowCenterWrap>
+  );
+};
 
-export default ProfilePage;
+export default withAuthentication(ProfilePage);
