+++ web/pages/profile/pwchange.js
@@ -2,8 +2,9 @@ import React from 'react';
 
 import PwChangeForm from '../../components/Forms/PwChangeForm';
 import * as GS from '../../components/GlobalStyle';
+import { withAuthentication } from '../../utils/with-authentication';
 
-const ProfilePage = () => {
+const PwChangePage = () => {
   return (
     <GS.FlexRowCenterWrap>
       <GS.SmallColumnBoard>
@@ -13,4 +14,4 @@ const ProfilePage = () => {
   );
 };
 
-export default ProfilePage;
+export default withAuthentication(PwChangePage);
